--- D-Chat/Models/TalkListModel.swift
@@ -1,11 +1,0 @@
-//
-//  TalkListModel.swift
-//  D-Chat
-//
-//  Created by Tanaka Soushi on 2022/12/03.
-//
-
-import Foundation
-
-final class TalkListModel {
-}
