--- D-Chat/Models/ViewDataModels/Friend.swift
@@ -1,13 +1,0 @@
-//
-//  Friend.swift
-//  D-Chat
-//
-//  Created by Tanaka Soushi on 2022/12/03.
-//
-
-import Foundation
-
-struct FriendViewDataModel {
-    let name: String
-    let talk: [TalkViewDataModel] = []
-}
