--- D-Chat/Models/ViewDataModels/Talk.swift
@@ -1,15 +1,0 @@
-//
-//  Talk.swift
-//  D-Chat
-//
-//  Created by Tanaka Soushi on 2022/12/03.
-//
-
-import Foundation
-
-struct TalkViewDataModel {
-    let name: String
-    let date: Date
-    let isOwn: Bool
-    let message: String
-}
