--- D-Chat/UIResource/Resource/Localizable/Localizable.strings
@@ -1,4 +1,0 @@
-
-// MARK: - サンプル
-"Friend" = "友達";
-"Talk" = "トーク";
