--- D-ChatUITests/D_ChatUITestsLaunchTests.swift
@@ -1,31 +1,0 @@
-//
-//  D_ChatUITestsLaunchTests.swift
-//  D-ChatUITests
-//
-//  Created by Tanaka Soushi on 2022/11/28.
-//
-
-import XCTest
-
-class DChatUITestsLaunchTests: XCTestCase {
-    override class var runsForEachTargetApplicationUIConfiguration: Bool {
-        true
-    }
-
-    override func setUpWithError() throws {
-        continueAfterFailure = false
-    }
-
-    func testLaunch() throws {
-        let app = XCUIApplication()
-        app.launch()
-
-        // Insert steps here to perform after app launch but before taking a screenshot,
-        // such as logging into a test account or navigating somewhere in the app
-
-        let attachment = XCTAttachment(screenshot: app.screenshot())
-        attachment.name = "Launch Screen"
-        attachment.lifetime = .keepAlways
-        add(attachment)
-    }
-}
