+++ Gourmet/Models/FavoriteGourmetListModel.swift
@@ -1,0 +1,11 @@
+//
+//  FavoriteGourmetListModel.swift
+//  Gourmet
+//
+//  Created by Tanaka Soushi on 2022/12/03.
+//
+
+import Foundation
+
+final class FavoriteGourmetListModel {
+}
