+++ Gourmet/Models/ViewDataModels/GourmetViewDataModel.swift
@@ -1,0 +1,12 @@
+//
+//  GourmetViewDataModel.swift
+//  Gourmet
+//
+//  Created by Tanaka Soushi on 2022/12/04.
+//
+
+import Foundation
+
+struct GourmetViewDataModel {
+    let name: String
+}
