+++ Gourmet/UIResource/Resource/Assets.xcassets/Text/Contents.json
