+++ Gourmet/UIResource/Resource/Assets.xcassets/Text/TextWhite.colorset/Contents.json
