+++ Gourmet/UIResource/Resource/Assets.xcassets/View/AccentPrimary.colorset/Contents.json
