+++ Gourmet/UIResource/Resource/Assets.xcassets/View/AccentSubPrimary.colorset/Contents.json
