+++ Gourmet/UIResource/Resource/Localizable/Localizable.strings
@@ -1,0 +1,4 @@
+
+// MARK: - サンプル
+"Search" = "検索";
+"Favorite" = "お気に入り";
