+++ GourmetUITests/GourmetUITestsLaunchTests.swift
@@ -1,0 +1,31 @@
+//
+//  D_ChatUITestsLaunchTests.swift
+//  GourmetUITests
+//
+//  Created by Tanaka Soushi on 2022/11/28.
+//
+
+import XCTest
+
+class GourmetUITestsLaunchTests: XCTestCase {
+    override class var runsForEachTargetApplicationUIConfiguration: Bool {
+        true
+    }
+
+    override func setUpWithError() throws {
+        continueAfterFailure = false
+    }
+
+    func testLaunch() throws {
+        let app = XCUIApplication()
+        app.launch()
+
+        // Insert steps here to perform after app launch but before taking a screenshot,
+        // such as logging into a test account or navigating somewhere in the app
+
+        let attachment = XCTAttachment(screenshot: app.screenshot())
+        attachment.name = "Launch Screen"
+        attachment.lifetime = .keepAlways
+        add(attachment)
+    }
+}
