+++ Podfile
@@ -1,20 +1,20 @@
 # Uncomment the next line to define a global platform for your project
 platform :ios, '14.0'
 
-target 'D-Chat' do
+target 'Gourmet' do
   # Comment the next line if you don't want to use dynamic frameworks
   use_frameworks!
 
-  # Pods for D-Chat
+  # Pods for Gourmet
   pod 'SwiftGen', '~> 6.6.2'
   pod 'SwiftLint', '~> 0.50.1'
   
-  target 'D-ChatTests' do
+  target 'GourmetTests' do
     inherit! :search_paths
     # Pods for testing
   end
 
-  target 'D-ChatUITests' do
+  target 'GourmetUITests' do
     # Pods for testing
   end
